maysol-back/requirements.txt
astroid==2.2.5
autopep8==1.4
backcall==0.1.0
blessings==1.7
bpython==0.18
certifi==2019.3.9
chardet==3.0.4
curtsies==0.3.0
decorator==4.3.0
dj-database-url==0.4.1
dj-static==0.0.6
Django==2.2
django-cors-headers==2.5.2
django-extensions==2.1.6
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-relativedelta==1.0.5
django-rest-auth==0.9.5
django-wkhtmltopdf==3.2.0
djangorestframework==3.9.2
et-xmlfile==1.0.1
#freeze==1.0.10
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
ipython==7.0.1
ipython-genutils==0.2.0
isort==4.3.17
jdcal==1.4
jedi==0.13.1
lazy-object-proxy==1.3.1
lxml==4.3.0
mccabe==0.6.1
mysqlclient==1.4.2.post1
openpyxl==2.5.12
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
ptyprocess==0.6.0
pycodestyle==2.4.0
Pygments==2.3.1
pylint==2.3.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.21.0
simplegeneric==0.8.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
traitlets==4.3.2
typed-ast==1.3.1
typing==3.6.6
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.15.2
whitenoise==4.1.2
wrapt==1.11.1
